@gyuho 2015-11-20
```
BenchmarkWatchableStoreSyncedCancel:
[1]:
benchmark                                 old ns/op       new ns/op     delta
BenchmarkWatchableStoreSyncedCancel       28481490911     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-2     24699111926     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-4     25444312207     0.07          -100.00%
BenchmarkWatchableStoreSyncedCancel-8     25213736765     0.06          -100.00%

benchmark                                 old allocs     new allocs     delta
BenchmarkWatchableStoreSyncedCancel       5001           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2     6133           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-4     7427           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-8     8278           0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkWatchableStoreSyncedCancel       2002296       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2     1997224       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-4     2126552       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-8     2165512       0             -100.00%


[2]:
benchmark                                 old ns/op       new ns/op     delta
BenchmarkWatchableStoreSyncedCancel       28405861525     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-2     24189639392     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-4     25185842756     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-8     24981059078     0.06          -100.00%

benchmark                                 old allocs     new allocs     delta
BenchmarkWatchableStoreSyncedCancel       5039           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2     5977           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-4     7258           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-8     8195           0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkWatchableStoreSyncedCancel       2020232       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2     1948480       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-4     2092416       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-8     2144696       0             -100.00%


[3]:
benchmark                                 old ns/op       new ns/op     delta
BenchmarkWatchableStoreSyncedCancel       28649823895     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-2     24409337089     0.07          -100.00%
BenchmarkWatchableStoreSyncedCancel-4     25467552651     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-8     25205423967     0.05          -100.00%

benchmark                                 old allocs     new allocs     delta
BenchmarkWatchableStoreSyncedCancel       5020           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2     6069           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-4     7378           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-8     8310           0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkWatchableStoreSyncedCancel       2011264       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2     1977624       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-4     2131168       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-8     2175312       0             -100.00%


[4]:
benchmark                                 old ns/op       new ns/op     delta
BenchmarkWatchableStoreSyncedCancel       28946238621     0.10          -100.00%
BenchmarkWatchableStoreSyncedCancel-2     24380060493     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-4     25614712892     0.05          -100.00%
BenchmarkWatchableStoreSyncedCancel-8     24995822673     0.06          -100.00%

benchmark                                 old allocs     new allocs     delta
BenchmarkWatchableStoreSyncedCancel       4993           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2     6077           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-4     7357           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-8     8298           0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkWatchableStoreSyncedCancel       2001784       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2     1978072       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-4     2114320       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-8     2166792       0             -100.00%


[5]:
benchmark                                 old ns/op       new ns/op     delta
BenchmarkWatchableStoreSyncedCancel       28442078021     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-2     24422697590     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-4     25397499735     0.06          -100.00%
BenchmarkWatchableStoreSyncedCancel-8     25589699437     0.06          -100.00%

benchmark                                 old allocs     new allocs     delta
BenchmarkWatchableStoreSyncedCancel       4978           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-2     6006           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-4     7317           0              -100.00%
BenchmarkWatchableStoreSyncedCancel-8     8427           0              -100.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkWatchableStoreSyncedCancel       1993072       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-2     1958024       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-4     2111760       0             -100.00%
BenchmarkWatchableStoreSyncedCancel-8     2190552       0             -100.00%
```

And the script that I used to measure the 1 million watcher performance:

```
package storage

import (
	"fmt"
	"io/ioutil"
	"log"
	"math/rand"
	"os"
	"path"
	"time"
)

func TEST() {
	tmpDir, err := ioutil.TempDir(os.TempDir(), "etcd_test_storage")
	if err != nil {
		log.Fatal(err)
	}
	tmpPath := path.Join(tmpDir, "database")
	s := newWatchableStore(tmpPath)
	defer func() {
		s.store.Close()
		os.Remove(tmpPath)
	}()

	// Put a key so that we can spawn watchers on that key
	testKey := []byte("foo")
	testValue := []byte("bar")
	s.Put(testKey, testValue)

	w := s.NewWatcher()

	// put 1 million watchers on the same key
	const watcherSize = 1000000

	cancels := make([]CancelFunc, watcherSize)
	for i := 0; i < watcherSize; i++ {
		// 0 for startRev to keep watchers in synced
		cancel := w.Watch(testKey, true, 0)
		cancels[i] = cancel
	}

	// randomly cancel watchers to make it not biased towards
	// data structures with an order, such as slice.
	ix := rand.Perm(watcherSize)

	ss := time.Now()
	for i, idx := range ix {
		if i%10000 == 0 {
			log.Println("Cancelling", i)
		}
		cancels[idx]()
	}
	fmt.Println("took:", time.Since(ss))
}
```

The whole workflow can be found: https://github.com/gyuho/etcd-scripts/tree/master/%233859_synced_map

Thanks,