@gyuho 2015-11-20
Thanks for feedback! I will push updates as you suggest shortly.