@gyuho 2015-11-20
@xiang90 Please review again. I addressed all the issue you mentions.
Performance stays the same. I ran the script on 1 million keys
and map implementation is much faster (9 min vs 400 ms). And
added some test cases.