@gyuho 2015-11-20
@xiang90 Just changed it to `log.Panicf`. Please take a look again. Thanks,